feat: 🎸 move lazy loading to routes
