+++ src/containers/not-found/index.js
@@ -9,7 +9,7 @@ const NotFound = () => (
     <Helmet title='Carbon - 404 Not Found!' />
     <h1>This is not the route you&apos;re looking for.</h1>
     <p>This is not the route I&apos;m looking for...</p>
-    <img src={NotFoundSVG} />
+    <img alt='404 Not Found' src={NotFoundSVG} />
   </div>
 );
